Change financial amounts to decimal

+++ Source/1BrokerApi.Net/ClientTools/IPositions.cs
@@ -13,6 +13,6 @@ namespace Jojatekok.OneBrokerAPI.ClientTools
         /// <param name="isClosableWithMarketOrder">Set this parameter to true if the position should be closed with a market order.</param>
         /// <param name="stopLoss">Stop loss for the position, once opened.</param>
         /// <param name="takeProfit">Take profit for the position, once opened.</param>
-        void EditPosition(ulong id, bool isClosableWithMarketOrder = false, double? stopLoss = null, double? takeProfit = null);
+        void EditPosition(ulong id, bool isClosableWithMarketOrder = false, decimal? stopLoss = null, decimal? takeProfit = null);
     }
 }
