+++ Source/1BrokerApi.Net/JsonObjects/Bar.cs
@@ -9,16 +9,16 @@ namespace Jojatekok.OneBrokerAPI.JsonObjects
         public string Time { get; private set; }
 
         [JsonProperty("o")]
-        public double o { get; private set; }
+        public decimal o { get; private set; }
 
         [JsonProperty("h")]
-        public double h { get; private set; }
+        public decimal h { get; private set; }
 
         [JsonProperty("l")]
-        public double l { get; private set; }
+        public decimal l { get; private set; }
 
         [JsonProperty("c")]
-        public double c { get;  private set; }
+        public decimal c { get;  private set; }
         
     }
 }
